+++ src/services/api/api.service.js
@@ -162,7 +162,12 @@ const updateProfileImages = ({ credentials, avatar = null, banner = null, backgr
     body: form
   })
     .then((data) => data.json())
-    .then((data) => parseUser(data))
+    .then((data) => {
+      if (data.error) {
+        throw new Error(data.error)
+      }
+      return parseUser(data)
+    })
 }
 
 const updateProfile = ({ credentials, params }) => {